feat: separate icons and documentation from constants
